Fix part 17

+++ math/0x02-calculus/17-integrate.py
@@ -17,6 +17,6 @@ def poly_integral(poly, C=0):
     result = []
     result.append(C)
     for x in range(len(poly)):
-        result.append(poly[x] / (x + 1))
+        result.append(int(poly[x] / (x + 1)))
 
     return result
